+++ jest.config.mjs
@@ -1,6 +1,11 @@
 export default {
   testEnvironment: 'node',
   transform: {
-    '^.+\.mjs$': 'babel-jest',
+    '^.+\\.mjs$': 'babel-jest',
   },
+  moduleFileExtensions: ['js', 'mjs'],
+  testMatch: ['**/__tests__/**/*.mjs'],
+  transformIgnorePatterns: [
+    'node_modules/(?!(@inquirer)/)',
+  ],
 };